Crate srml_executive

Executive Module

The Executive module acts as the orchestration layer for the runtime. It dispatches incoming extrinsic calls to the respective modules in the runtime.

Overview

The executive module is not a typical SRML module providing functionality around a specific feature. It is a cross-cutting framework component for the SRML. It works in conjunction with the SRML System module to perform these cross-cutting functions.

The Executive module provides functions to:

Check transaction validity.

Initialize a block.

Apply extrinsics.

Execute a block.

Finalize a block.

Start an off-chain worker.

Implementations

The Executive module provides the following implementations:

ExecuteBlock: Trait that can be used to execute a block.

Executive: Type that can be used to make the SRML available from the runtime.

Usage

The default Substrate node template declares the Executive type in its library.

Example

Executive type declaration from the node template.

/// Executive: handles dispatch to the various modules.
pub type Executive = executive::Executive<Runtime, Block, Context, Balances, Runtime, AllModules>;

Structs

Executive

Traits

ExecuteBlock
Trait that can be used to execute a block.

Type Definitions

CallOf

CheckedOf

OriginOf